
\name{crlag}
\alias{crlag}
\description{Apply the Circular Shift Operator to a Vector}
\title{Apply the Circular Shift Operator to a Vector}

\usage{crlag(x)}

\arguments{
\item{x}{An array.}
}
\value{
\item{crlag}{returns the array that results from applying 
the circular shift operator to the original array.}
}
