
\name{freqs}
\alias{freqs}
\description{Form Vector of Natural (Nyquist) Frequencies}
\title{Form Vector of Natural (Nyquist) Frequencies}

\usage{freqs(n)}

\arguments{
\item{n}{Integer indicating the number of desired frequencies.}
}
\value{
\item{freqs}{returns an array of length {\code{n}} of 
frequencies all contained in the interval [0,0.5].}
}
