
\name{infqnt}
\alias{infqnt}
\description{Plot Informative Quantile Function of a Data Set}
\title{Plot Informative Quantile Function of a Data Set}

\usage{infqnt(x)}

\arguments{
\item{x}{Array of length $n$ containing the data.}
}

\value{
\item{infqnt}{returns a plot of the informative quantile 
function for the data set {\code{x}}.}
}
