

\name{masmooth}
\alias{masmooth}
\description{Apply Moving Average Smoother to a Vector (S Version)}
\title{Apply Moving Average Smoother to a Vector (S Version)}

\usage{masmooth(x,k)}

\arguments{
\item{x}{Array containing the data.}
\item{k}{Length of the moving average smoother.}
}

\value{
\item{masmooth}{returns the resulting smoothed array 
{\code{z}}.}
}
