
\name{movave}
\alias{movave}
\description{Apply Moving Average Smoother to a Vector (Fortran Version)}
\title{Apply Moving Average Smoother to a Vector (Fortran Version)}

\usage{movave(x,k)}

\arguments{
\item{x}{Array containing the data.}
\item{k}{Length of the moving average smoother.}
}

\value{
\item{movave}{returns the resulting smoothed array {\code{z}}.}
}
