
\name{movord}
\alias{movord}
\description{Apply Moving Order Statistic Operator to a Vector}
\title{Apply Moving Order Statistic Operator to a Vector}

\usage{movord(x,nord,k)}

\arguments{
\item{x}{Array of length $n$ containing the data.}
\item{nord}{Order of moving order statistic desired.}
\item{k}{Length of moving average smoother.}
}

\value{
\item{movord}{returns the array of moving order 
statistics.}
}
