
\name{multpoly}
\alias{multpoly}
\description{Multiply Two Polynomials}
\title{Multiply Two Polynomials}

\usage{multpoly(alpha,beta)}

\arguments{
\item{alpha}{Array of length $p$ containing the coefficients of the first 
polynomial.}
\item{beta}{Array of length $q$ containing the coefficients of the second 
polynomial.}
}
\value{
\item{multpoly}{returns an array of length $p+q$ 
containing the coefficients of the product of the polynomials.}
}
