
\name{odot}
\alias{odot}
\description{Form Outer Product of Two Vectors}
\title{Form Outer Product of Two Vectors}
\usage{odot(x,y)}

\arguments{
\item{x}{A vector.}
\item{y}{A vector.}
}

\value{
\item{odot}{returns an $n\times n$ matrix that is the outer product
of the vectors {\code{x}} and {\code{y}}.}
}
