
\name{partar}
\alias{partar}
\description{Calculate AR Coefficients from Autocorrelations}
\title{Calculate AR Coefficients from Autocorrelations}

\usage{partar(theta)}

\arguments{
\item{theta}{Array of length $p$ containing partial autocorrelations.}
}
\value{
\item{partar}{returns an array of length $p$ containing 
AR coefficients.}
}
