

\name{perdgm}
\alias{perdgm}
\description{Calculate Periodogram of a Time Series}
\title{Calculate Periodogram of a Time Series}

\usage{perdgm(x)}

\arguments{
\item{x}{Array of length $n$ containing the data.}
}
\value{
\item{perdgm}{returns the periodogram of {\code{x}} at 
the frequencies $j/n,j=0,1,\ldots,[n/2]$.}
}
