
\name{rtpoly}
\alias{rtpoly}
\description{Find the Coefficients of a Polynomial Given its Roots}
\title{Find the Coefficients of a Polynomial Given its Roots}

\usage{rtpoly(roots)}

\arguments{
\item{roots}{Array of length $p$ containing the zeros of the polynomial.}
}

\value{
\item{rtpoly}{returns an array of length $p$ 
containing the coefficients of the polynomial with the given roots.}
}
