
\name{rw}
\alias{rw}
\description{Simulate Data from a Random Walk Process}
\title{Simulate Data from a Random Walk Process}

\usage{rw(n,seed=0)}

\arguments{
\item{n}{Integer containing the length of the desired realization.}
\item{seed}{Real scalar containing the seed for the random number generator.}
}

\value{
\item{rw}{returns an array of length {\code{n}} containing
a realization of a Gaussian random walk.}
}
