
\name{schur}
\alias{schur}
\description{Form the Schur Matrix Corresponding to AR Parameters}
\title{Form the Schur Matrix Corresponding to AR Parameters}

\usage{schur(alpha)}

\arguments{
\item{alpha}{An array of length $p$ containing the AR coefficients 
$\Valpha$.}
}
\value{
\item{schur}{returns the Schur matrix for the AR coefficients.}
}
