

\name{simcfs}
\alias{simcfs}
\description{Simulate Coefficients of a Polynomial with Zeros Outside
\title{Simulate Coeff. of a Polynomial with Zeros Outside
Unit Circle}

\usage{simcfs(p,seed=0)}

\arguments{
\item{p}{Degree of desired polynomial.}
\item{seed}{Real scalar containing the seed for the random number generator.}
}

\value{
\item{simcfs}{returns an array containing coefficients
of a polynomial having all of its zeros greater than one in modulus.}
}
