
\name{stdf}
\alias{stdf}
\description{Standardize a Vector}
\title{Standardize a Vector}

\usage{stdf(f,fac,a,b)}

\arguments{
\item{f}{A vector.}
\item{fac}{A nonzero real scalar to divide {\code{f}} by.}
\item{a}{Real scalar indicating lower limit allowed for standardized 
${\tt{f}}/{\tt{fac}}$.}
\item{b}{Real scalar indicating upper limit allowed for standardized
${\tt{f}}/{\tt{fac}}$. }
}

\value{
\item{stdf}{returns the standardized vector {\code{f}}.}
}
