
\name{wntest}
\alias{wntest}
\description{Form Plots for White Noise Test}
\title{Form Plots for White Noise Test}

\usage{wntest(x,m,alpha=0.05)}

\arguments{
\item{x}{An array containing the realization to be tested.}
\item{m}{Integer containing the number of correlations.}
\item{alpha}{Real scalar ($0<${\code{alpha}}$<1$) indicating the 
level of confidence at which the test is to be preformed.}
}

\value{
\item{wntest}{returns plots of two white noise test for time series
data.}
}

